moved \Drupal\Core\Plugin\ResourceBase
