+++ core/modules/dblog/src/Plugin/rest/resource/DBLogResource.php
@@ -2,7 +2,7 @@
 
 namespace Drupal\dblog\Plugin\rest\resource;
 
-use Drupal\rest\Plugin\ResourceBase;
+use Drupal\Core\Plugin\ResourceBase;
 use Drupal\rest\ResourceResponse;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
